Keep all issues populating logic in Job
